bug fixes
Added some additional safety checks. Cleaned up static data not in use. Improved a few of the protection checks. Fixed an issue with https. improved credential safety. + + +...

+++ development/cycle_record.h
@@ -15,12 +15,22 @@ struct CycleRecord {
   uint16_t end_v_dv;        // pack voltage dV at DONE entry
   uint8_t  start_soc;       // BMS SOC at CC entry  (255 = not available)
   uint8_t  end_soc;         // BMS SOC at DONE entry (255 = not available)
-  int8_t   start_temp;      // monolithMaxTemp (°C) at CC entry
-  int8_t   end_temp;        // monolithMaxTemp (°C) at DONE entry
-  uint16_t total_ah_x100;   // session.chargeAh × 100  (2 dp fixed-point)
-  uint32_t total_wh_x10;    // session.energyWh × 10   (1 dp fixed-point)
+  int8_t   start_temp;      // monolithMaxTemp (°C) at CC entry; -128 = no valid reading
+  int8_t   end_temp;        // monolithMaxTemp (°C) at DONE entry; -128 = no valid reading
+  // Per-cycle totals (this cycle only), NOT the session-to-date counters:
+  // rampTask totals them tick by tick from CC entry to DONE (cycleAh/cycleWh),
+  // never derived from the session accumulators, so Reset Session cannot
+  // affect them. uint32_t, not uint16_t — at x100 a uint16_t wrapped silently
+  // past 655.35 Ah.
+  uint32_t total_ah_x100;   // Ah delivered this cycle × 100  (2 dp fixed-point)
+  uint32_t total_wh_x10;    // Wh delivered this cycle × 10   (1 dp fixed-point)
   uint16_t bulk_min;        // minutes in PHASE_CC
   uint16_t absorption_min;  // minutes in PHASE_CV
   uint8_t  charger_count;   // ctrl.chargerCount at CC entry
+  // How absorption ended. 0 is the only "ran to completion" value; 1 and 2 both
+  // mean absorption was cut short, for different reasons, and Stage 2 training
+  // should weight them differently from a natural finish.
   uint8_t  abort_reason;    // 0 = current_taper   1 = cv_timeout
+                            // 2 = above_target (pack went above the ceiling, or
+                            //     the ceiling was lowered below the pack)
 };
